Update non-interactive proof testing
