+++ tests/Tests/NonInteractiveProof.hs
@@ -7,12 +7,11 @@ import           Data.Typeable                               (Proxy (..), Typeab
 import           Prelude                                     hiding (Fractional (..), Num (..), length)
 import           Test.Hspec                                  (describe, hspec, it)
 import           Test.QuickCheck                             (Arbitrary, Testable (property), withMaxSuccess)
-import           Tests.NonInteractiveProof.Internal          (NonInteractiveProofTestData (..))
 import           Tests.NonInteractiveProof.Plonk             (PlonkBS, specPlonk)
 
 import           ZkFold.Base.Algebra.EllipticCurve.BLS12_381
 import           ZkFold.Base.Protocol.Commitment.KZG         (KZG)
-import           ZkFold.Base.Protocol.NonInteractiveProof    (NonInteractiveProof (..))
+import           ZkFold.Base.Protocol.NonInteractiveProof    (NonInteractiveProof (..), NonInteractiveProofTestData(..))
 
 propNonInteractiveProof :: forall a .
     NonInteractiveProof a =>
