+++ zkfold-base.cabal
@@ -125,7 +125,9 @@ library
       ZkFold.Base.Protocol.ARK.Protostar.SpecialSound
       ZkFold.Base.Protocol.Commitment.KZG
       ZkFold.Base.Protocol.NonInteractiveProof
+      ZkFold.Base.Protocol.NonInteractiveProof.Internal
       ZkFold.Base.Protocol.NonInteractiveProof.Prover
+      ZkFold.Base.Protocol.NonInteractiveProof.Testing
       ZkFold.Prelude
       ZkFold.Symbolic.Algorithms.Hash.Blake2b
       ZkFold.Symbolic.Algorithms.Hash.Blake2b.Constants
@@ -249,7 +251,6 @@ test-suite zkfold-base-test
       Tests.GroebnerBasis
       Tests.Group
       Tests.NonInteractiveProof
-      Tests.NonInteractiveProof.Internal
       Tests.NonInteractiveProof.Plonk
       Tests.Pairing
       Tests.Permutations
